refactor(cloudkit): split RecordMapping per record type (460→≤88) (#272, #275, #267, #279)
Move each `extension X: CloudKitRecordConvertible` conformance into its
own sibling `<Record>+CloudKit.swift` file. The core protocols
(CloudKitRecordConvertible, IdentifiableRecord, SystemFieldsCacheable),
the CKRecord encodedSystemFields extension, and RecordTypeRegistry stay
in a renamed core file CloudKitRecordConvertible.swift.

Drops four baseline entries entirely:
- file_length (460 lines) — largest new file is 88.
- file_name — core file name matches its primary protocol; per-record
  files match the Type+Suffix convention.
- cyclomatic_complexity on TransactionRecord.toCKRecord — extracted
  the importOrigin branches into encodeImportOriginFields(into:); same
  fields, same order, same CKRecord mutations.
- identifier_name on 14 `if let v =` shadow bindings — renamed to
  `if let value =`.

No behaviour change. No access widenings (everything crossing the split
was already internal). project.yml uses a Backends glob so new files
are picked up automatically.
